ft_printf3

+++ ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: tibarike <tibarike@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/21 20:36:51 by tibarike          #+#    #+#             */
-/*   Updated: 2024/11/23 15:38:25 by tibarike         ###   ########.fr       */
+/*   Updated: 2024/12/02 13:52:53 by tibarike         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-int	ft_strlen(const char *str)
+size_t	ft_strlen(const char *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i])
